feat(copyWriting): refactor Template and update template data
This commit refactors the Template interface into a type and introduces
an enum for TemplateType. It also updates the templates data to use the
new type and enum. Additionally, it optimizes the rendering of output
messages by using dangerouslySetInnerHTML. This is done to allow for the
rendering of HTML elements within the output messages. #lotte