feat(npm): add new connect SID button component
This commit adds a new connect SID button component to the npm package and modifies the package.json to point to the newly added component. The rollup.config.js and tsconfig.json files have also been added to aid in the building process. The version of the package is now bumped to 1.0.0-alpha.4.

NOTE: The new component is composed of the `SIDButton` and `SidSVG` components, with the `SIDButton` being a more complex component that depends on the state of the `isConnected` prop to render different button styles and functionality. The button styles and functionality are all defined inside the `SIDButton` component.

Resolves: #1234