feat(chat): update chat box and message styling (#feature-chat-new-design)
Updated the chat box and message styling in the chat application. The changes include adjustments to padding, borders, and color schemes in the ChatBox.module.scss, ChatMessage.module.scss, and Home.module.scss files. Avatar styling was added to messages in the ChatMessage.module.scss file. Adjustments were also made to the index.tsx file to add a class to the SID button, and changes were made in the ChatBox.tsx component to improve message handling. Additionally, a minor adjustment was made to the global.scss file to modify the body margin.