feat: Improve iframe resizing mechanism in index.tsx
This commit enhances the iframe resizing mechanism in the index.tsx file.
Instead of directly calling the resizeIframeToContentSize function, a
postMessage is now sent to the parent window with the new height of the
container. This change provides more flexibility and control over how
the resizing is handled. Also, a helpful console log message has been added
to assist with debugging and to provide more transparency about what the
code is doing.